Keep the cart after placing an order from the product page

+++ app/code/community/Bolt/Boltpay/Model/Observer.php
@@ -112,13 +112,20 @@ class Bolt_Boltpay_Model_Observer
     }
 
     /**
-     * Clears the Shopping Cart after the success page
+     * Clears the Shopping Cart except product page checkout order after the success page
      *
      * Event: checkout_onepage_controller_success_action
+     * @param $observer
      */
-    public function clearShoppingCart() {
+    public function clearShoppingCartExceptPPCOrder()
+    {
         $cartHelper = Mage::helper('checkout/cart');
-        $cartHelper->getCart()->truncate()->save();
+        if (Mage::app()->getRequest()->getParam('checkoutType') == Bolt_Boltpay_Block_Checkout_Boltpay::CHECKOUT_TYPE_PRODUCT_PAGE) {
+            $quoteId = Mage::app()->getRequest()->getParam('session_quote_id');
+            Mage::getSingleton('checkout/session')->setQuoteId($quoteId);
+        } else {
+            $cartHelper->getCart()->truncate()->save();
+        }
     }
 
     public function sendCompleteAuthorizeRequest($request)
